// src/enums/country.ts
export enum Country {
  Mexico = 'MX',
  Canada = 'CA',
  Spain = 'ES',
  Brazil = 'BR',
  Germany = 'DE',
  France = 'FR',
  Uruguay = 'UY',
  Italy = 'IT',
  Argentina = 'AR',
  Australia = 'AU',
  Poland = 'PL',
  Japan = 'JP',
  SouthKorea = 'KR',
  Netherlands = 'NL',
  Belgium = 'BE',
  Portugal = 'PT',
  Sweden = 'SE',
  Switzerland = 'CH',
  Denmark = 'DK',
  Norway = 'NO',
}